Move all resources to where they are used
