+++ src/App/app.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "../variables";
 
 .core19-app {
     background-color: $gray-50;
