+++ src/App/logo.scss
