+++ src/App/nav-bar.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "../variables";
 
 .core19-nav-bar {
     width: 100%;
