+++ src/App/nordic-logo-blue-icon-only.png
