+++ src/App/nordic-logo-gray-icon-only.png
