+++ src/App/nordic-logo-white-icon-only.png
