+++ src/App/visibility-bar.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "../variables";
 
 .core19-visibility-bar {
     display: flex;
