+++ src/Device/device-selector.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "../variables";
 
 .core19-device-selector {
     .dropdown-toggle.btn {
