+++ src/ErrorDialog/error.scss
