+++ src/Log/LogEntry.jsx
@@ -39,7 +39,7 @@ import { shape, number, string } from 'prop-types';
 import moment from 'moment';
 import { shell } from 'electron';
 
-import '../../resources/css/log-entry.scss';
+import './log-entry.scss';
 
 const regex = /(.*?)(https?:\/\/[^\s]+)/g;
