+++ src/Log/log-entry.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "../variables";
 
 .core19-log-entry {
     height: 20px;
