+++ src/Log/log-header.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "../variables";
 
 .core19-log-header {
     display: flex;
