+++ src/Log/log-viewer.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "../variables";
 
 .core19-log-viewer {
     background-color: $gray-800;
